backend/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
django-unfold==0.63.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
django-redis==5.4.0
pillow==11.3.0
PyJWT==2.10.1
sqlparse==0.5.3
celery==5.3.4
redis==5.0.1
